+++ lib/umya_native.ex
@@ -2191,5 +2191,41 @@ defmodule UmyaNative do
   @spec test_simple_function() :: String.t()
   def test_simple_function(), do: error()
 
+  # VML Shape getter functions
+  @spec get_vml_shape_style(reference(), String.t(), String.t()) ::
+          {:ok, String.t()} | {:error, atom()}
+  def get_vml_shape_style(_spreadsheet, _sheet_name, _shape_id),
+    do: error()
+
+  @spec get_vml_shape_type(reference(), String.t(), String.t()) ::
+          {:ok, String.t()} | {:error, atom()}
+  def get_vml_shape_type(_spreadsheet, _sheet_name, _shape_id),
+    do: error()
+
+  @spec get_vml_shape_filled(reference(), String.t(), String.t()) ::
+          {:ok, boolean()} | {:error, atom()}
+  def get_vml_shape_filled(_spreadsheet, _sheet_name, _shape_id),
+    do: error()
+
+  @spec get_vml_shape_fill_color(reference(), String.t(), String.t()) ::
+          {:ok, String.t()} | {:error, atom()}
+  def get_vml_shape_fill_color(_spreadsheet, _sheet_name, _shape_id),
+    do: error()
+
+  @spec get_vml_shape_stroked(reference(), String.t(), String.t()) ::
+          {:ok, boolean()} | {:error, atom()}
+  def get_vml_shape_stroked(_spreadsheet, _sheet_name, _shape_id),
+    do: error()
+
+  @spec get_vml_shape_stroke_color(reference(), String.t(), String.t()) ::
+          {:ok, String.t()} | {:error, atom()}
+  def get_vml_shape_stroke_color(_spreadsheet, _sheet_name, _shape_id),
+    do: error()
+
+  @spec get_vml_shape_stroke_weight(reference(), String.t(), String.t()) ::
+          {:ok, String.t()} | {:error, atom()}
+  def get_vml_shape_stroke_weight(_spreadsheet, _sheet_name, _shape_id),
+    do: error()
+
   defp error(), do: :erlang.nif_error(:nif_not_loaded)
 end