+++ lib/umya_native/vml_support.ex
@@ -40,4 +40,40 @@ defmodule UmyaNative.VmlSupport do
           :ok | {:error, atom()}
   def set_vml_shape_stroke_weight(_spreadsheet, _sheet_name, _shape_id, _stroke_weight),
     do: :erlang.nif_error(:nif_not_loaded)
+
+  # Getter functions for VML shape properties
+  @spec get_vml_shape_style(reference(), String.t(), String.t()) ::
+          {:ok, String.t()} | {:error, atom()}
+  def get_vml_shape_style(_spreadsheet, _sheet_name, _shape_id),
+    do: :erlang.nif_error(:nif_not_loaded)
+
+  @spec get_vml_shape_type(reference(), String.t(), String.t()) ::
+          {:ok, String.t()} | {:error, atom()}
+  def get_vml_shape_type(_spreadsheet, _sheet_name, _shape_id),
+    do: :erlang.nif_error(:nif_not_loaded)
+
+  @spec get_vml_shape_filled(reference(), String.t(), String.t()) ::
+          {:ok, boolean()} | {:error, atom()}
+  def get_vml_shape_filled(_spreadsheet, _sheet_name, _shape_id),
+    do: :erlang.nif_error(:nif_not_loaded)
+
+  @spec get_vml_shape_fill_color(reference(), String.t(), String.t()) ::
+          {:ok, String.t()} | {:error, atom()}
+  def get_vml_shape_fill_color(_spreadsheet, _sheet_name, _shape_id),
+    do: :erlang.nif_error(:nif_not_loaded)
+
+  @spec get_vml_shape_stroked(reference(), String.t(), String.t()) ::
+          {:ok, boolean()} | {:error, atom()}
+  def get_vml_shape_stroked(_spreadsheet, _sheet_name, _shape_id),
+    do: :erlang.nif_error(:nif_not_loaded)
+
+  @spec get_vml_shape_stroke_color(reference(), String.t(), String.t()) ::
+          {:ok, String.t()} | {:error, atom()}
+  def get_vml_shape_stroke_color(_spreadsheet, _sheet_name, _shape_id),
+    do: :erlang.nif_error(:nif_not_loaded)
+
+  @spec get_vml_shape_stroke_weight(reference(), String.t(), String.t()) ::
+          {:ok, String.t()} | {:error, atom()}
+  def get_vml_shape_stroke_weight(_spreadsheet, _sheet_name, _shape_id),
+    do: :erlang.nif_error(:nif_not_loaded)
 end